Revert partial merge (commit ee46392).
